ml/requirements.txt
pandas==2.2.2
scikit-learn==1.5.0
psycopg2-binary>=2.9.11
joblib==1.4.2
pydantic-settings==2.6.1